Make heroku/nodejs-engine available as a library
